Rename directories and files in order to pass all of the tests.
