+++ controllers/locations.controller.js
@@ -2,7 +2,7 @@
 
 const express = require('express');
 const locations = express.Router();
-const locationsArray = require('../models/locations.model.js');
+const locationsArray = require('../models/location.model.js');
 
 locations.get('/', (req, res) => {
 	res.json(locationsArray);
