+++ controllers/machines.controller.js
@@ -2,7 +2,7 @@
 
 const express = require('express');
 const machines = express.Router();
-const machinesArray = require('../models/machines.model.js');
+const machinesArray = require('../models/machine.model.js');
 
 machines.get('/', (req, res) => {
 	res.json(machinesArray);
