+++ controllers/persons.controller.js
@@ -2,7 +2,7 @@
 
 const express = require('express');
 const persons = express.Router();
-const personsArray = require('../models/persons.model.js');
+const personsArray = require('../models/person.model.js');
 
 persons.get('/', (req, res) => {
 	res.json(personsArray);
