+++ controllers/plans.controller.js
@@ -2,7 +2,7 @@
 
 const express = require('express');
 const plans = express.Router();
-const plansArray = require('../models/plans.model.js');
+const plansArray = require('../models/plan.model.js');
 
 plans.get('/', (req, res) => {
 	res.json(plansArray);
