+++ controllers/special-events.controller.js
@@ -2,7 +2,7 @@
 
 const express = require('express');
 const specialEvents = express.Router();
-const specialEventsArray = require('../models/specialEvents.model.js');
+const specialEventsArray = require('../models/special-event.model.js');
 
 specialEvents.get('/', (req, res) => {
 	res.json(specialEventsArray);
